pyproject.toml
[project]
name = "fx-swap-bot"
version = "0.1.0"
description = "AI-powered FX Swap Arbitrage Bot"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.10",
    "openai>=1.50",
    "pandas>=2.2",
    "numpy>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.31"  # <--- 必須追加
]